Exceptions added

+++ JEAU1H_HFT_2021222.Logic/GameLogic.cs
@@ -36,12 +36,26 @@ namespace JEAU1H_HFT_2021222.Logic
 
         public void Create(Game item)
         {
-            this.repo.Create(item);
+            if (int.Parse(item.Pyear) > 1980)
+            {
+                this.repo.Create(item);
+            }
+            else
+            {
+                throw new ArgumentException("This game is too old");
+            }
         }
 
         public void Delete(int id)
         {
-            this.repo.Delete(id);
+            if (repo.Read(id) != null)
+            {
+                this.repo.Delete(id);
+            }
+            else
+            {
+                throw new ArgumentException("Not found");
+            }
         }
 
         public Game Read(int id)
